node: Ensure first pod CIDR is used for router IP
The setup of the primary Cilium router IP requires to know the CIDR to
which this IP belongs. In cluster pool IPAM mode, this will be the first
pod CIDR listed in the CiliumNode CRD. The current code assumes that
there is only one pod CIDR per IP version, but with dynamic cluster
pool, we will have potentially multiple pod CIDRs. Therefore, this
commit ensures that only the first pod CIDR is used as the node
addressing CIDR.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>